Refactor(component-name): extract useEffect and useState logic into a separate hook in RandomDish component
